исправила компонент App
